+++ rollup.config.js
@@ -61,11 +61,18 @@ const normalBundle = {
     }),
     scss({
       output: pkg.style,
+      prefix: `@import "./variables.scss";`,
     }),
-    copy([{ files: 'src/assets/*', dest: 'dist/assets' }], {
-      verbose: true,
-      watch: process.env.ROLLUP_WATCH,
-    }),
+    copy(
+      [
+        { files: 'src/assets/*', dest: 'dist/assets' },
+        { files: 'src/i18n/*.json', dest: 'dist/i18n' },
+      ],
+      {
+        verbose: true,
+        watch: process.env.ROLLUP_WATCH,
+      },
+    ),
     url(),
     commonjs(),
     json(),
@@ -123,6 +130,8 @@ const fullBrowserBundle = {
       process: true,
       buffer: true,
     }),
+    copy([{ files: 'src/i18n/*.json', dest: 'dist/i18n' }]),
+    // terser(),
   ],
 };