+++ src/components/ChannelPreviewLastMessage.js
@@ -2,6 +2,7 @@ import React, { PureComponent } from 'react';
 import { Avatar } from './Avatar';
 import PropTypes from 'prop-types';
 import truncate from 'lodash/truncate';
+import { withTranslationContext } from '../context';
 
 /**
  * Used as preview component for channel item in [ChannelList](#channellist) component.
@@ -10,7 +11,7 @@ import truncate from 'lodash/truncate';
  * @extends PureComponent
  */
 
-export class ChannelPreviewLastMessage extends PureComponent {
+class ChannelPreviewLastMessage extends PureComponent {
   static propTypes = {
     /** **Available from [chat context](https://getstream.github.io/stream-chat-react/#chat)** */
     channel: PropTypes.object.isRequired,
@@ -45,6 +46,8 @@ export class ChannelPreviewLastMessage extends PureComponent {
   };
 
   render() {
+    const { t } = this.props;
+
     const unreadClass =
       this.props.unread >= 1 ? 'str-chat__channel-preview--unread' : '';
     const activeClass = this.props.active
@@ -64,7 +67,7 @@ export class ChannelPreviewLastMessage extends PureComponent {
             <span className="str-chat__channel-preview-title">{name}</span>
             <span className="str-chat__channel-preview-last-message">
               {!this.props.channel.state.messages[0]
-                ? 'Nothing yet...'
+                ? t('Nothing yet...')
                 : truncate(this.props.latestMessage, {
                     length: this.props.latestMessageLength,
                   })}
@@ -80,3 +83,6 @@ export class ChannelPreviewLastMessage extends PureComponent {
     );
   }
 }
+
+ChannelPreviewLastMessage = withTranslationContext(ChannelPreviewLastMessage);
+export { ChannelPreviewLastMessage };