+++ src/components/ChatDown.tsx
@@ -4,14 +4,15 @@ import PropTypes from 'prop-types';
 import { LoadingChannels } from './LoadingChannels';
 
 import placeholder from '../assets/str-chat__connection-error.svg';
+import { withTranslationContext } from '../context';
 
 export interface ChatDownProps {
-    /** The image url for this error */
-    image: string;
-    /** The type of error */
-    type: string;
-    /** The error message to show */
-    text: string;
+  /** The image url for this error */
+  image: string;
+  /** The type of error */
+  type: string;
+  /** The error message to show */
+  text: string;
 }
 
 /**
@@ -20,7 +21,7 @@ export interface ChatDownProps {
  * @example ./docs/ChatDown.md
  * @extends PureComponent
  */
-export class ChatDown extends React.PureComponent<ChatDownProps> {
+class ChatDown extends React.PureComponent<ChatDownProps> {
   static propTypes = {
     /** The image url for this error */
     image: PropTypes.string,
@@ -33,11 +34,10 @@ export class ChatDown extends React.PureComponent<ChatDownProps> {
   static defaultProps = {
     image: placeholder,
     type: 'Error',
-    text: 'Error connecting to chat, refresh the page to try again.',
   };
 
   render() {
-    const { image, type, text } = this.props;
+    const { image, type, text, t } = this.props;
 
     return (
       <div className="str-chat__down">
@@ -45,9 +45,15 @@ export class ChatDown extends React.PureComponent<ChatDownProps> {
         <div className="str-chat__down-main">
           <img src={image} />
           <h1>{type}</h1>
-          <h3>{text}</h3>
+          <h3>
+            {text ||
+              t('Error connecting to chat, refresh the page to try again.')}
+          </h3>
         </div>
       </div>
     );
   }
 }
+
+ChatDown = withTranslationContext(ChatDown);
+export { ChatDown };