+++ src/components/DateSeparator.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Moment from 'moment';
 import PropTypes from 'prop-types';
+import { withTranslationContext } from '../context';
 
 /**
  * DateSeparator - A simple date separator
@@ -8,7 +8,7 @@ import PropTypes from 'prop-types';
  * @example ./docs/DateSeparator.md
  * @extends PureComponent
  */
-export class DateSeparator extends React.PureComponent {
+class DateSeparator extends React.PureComponent {
   static propTypes = {
     /** The date to format */
     date: PropTypes.instanceOf(Date),
@@ -22,7 +22,7 @@ export class DateSeparator extends React.PureComponent {
     position: 'right',
   };
   render() {
-    const { position } = this.props;
+    const { position, tDateTimeParser } = this.props;
     if (!Date.parse(this.props.date)) {
       return null;
     }
@@ -34,14 +34,7 @@ export class DateSeparator extends React.PureComponent {
         <div className="str-chat__date-separator-date">
           {this.props.formatDate
             ? this.props.formatDate(this.props.date)
-            : Moment(this.props.date.toISOString()).calendar(null, {
-                lastDay: '[Yesterday]',
-                sameDay: '[Today]',
-                nextDay: '[Tomorrow]',
-                lastWeek: '[Last] dddd',
-                nextWeek: 'dddd',
-                sameElse: 'L',
-              })}
+            : tDateTimeParser(this.props.date.toISOString()).calendar()}
         </div>
         {(position === 'left' || position === 'center') && (
           <hr className="str-chat__date-separator-line" />
@@ -50,3 +43,7 @@ export class DateSeparator extends React.PureComponent {
     );
   }
 }
+
+DateSeparator = withTranslationContext(DateSeparator);
+
+export { DateSeparator };