+++ src/components/EmptyStateIndicator.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { withTranslationContext } from '../context';
 
-export const EmptyStateIndicator = (listType: string) => {
+export const EmptyStateIndicator = (listType: string, t: any) => {
   let Indicator;
   switch (listType) {
     case 'channel':
-      Indicator = <p>You have no channels currently</p>;
+      Indicator = <p>{t('You have no channels currently')}</p>;
       break;
     case 'message':
       Indicator = null;