+++ src/components/MessageInput.js
@@ -130,6 +130,16 @@ class MessageInput extends PureComponent {
      * Defaults to and accepts same props as: [SendButton](https://getstream.github.io/stream-chat-react/#sendbutton)
      * */
     SendButton: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+    /**
+     * Any additional attrubutes that you may want to add for underlying HTML textarea element.
+     * e.g.
+     * <MessageInput
+     *  additionalTextareaProps={{
+     *    maxLength: 10,
+     *  }}
+     * />
+     */
+    additionalTextareaProps: PropTypes.object,
   };
 
   static defaultProps = {
@@ -139,6 +149,7 @@ class MessageInput extends PureComponent {
     maxRows: 10,
     Input: MessageInputLarge,
     SendButton,
+    additionalTextareaProps: {},
   };
 
   componentDidMount() {
@@ -276,9 +287,7 @@ class MessageInput extends PureComponent {
     const text = this.state.text;
     // the channel component handles the actual sending of the message
     const attachments = [...this.state.attachments];
-    if (this.props.message && this.props.message.attachments) {
-      attachments.push(...this.props.message.attachments);
-    }
+
     for (const id of this.state.imageOrder) {
       const image = this.state.imageUploads[id];
       if (!image || image.state === 'failed') {
@@ -307,6 +316,10 @@ class MessageInput extends PureComponent {
         // TODO: show error to user that they should wait until image is uploaded
         return;
       }
+      const dupe = attachments.filter(
+        (attach) => upload.asset_url === attach.url,
+      );
+      if (dupe.length >= 1) continue;
       attachments.push({
         type: 'file',
         asset_url: upload.url,
@@ -495,6 +508,7 @@ class MessageInput extends PureComponent {
 
     let response = {};
     response = {};
+
     try {
       if (this.props.doFileUploadRequest) {
         response = await this.props.doFileUploadRequest(